+++ pywiim/cli/diagnostics.py
@@ -406,11 +406,11 @@ class DeviceDiagnostics:
             elif "wiim mini" in model_lower:
                 available_modes = ["Line Out", "Optical Out"]
             elif "wiim ultra" in model_lower:
-                available_modes = ["Line Out", "Optical Out", "Coax Out", "Bluetooth Out", "Headphone Out", "HDMI Out"]
+                available_modes = ["Line Out", "Optical Out", "Coax Out", "Headphone Out", "HDMI Out"]
             elif "wiim pro" in model_lower or "wiim" in model_lower:
-                available_modes = ["Line Out", "Optical Out", "Coax Out", "Bluetooth Out"]
+                available_modes = ["Line Out", "Optical Out", "Coax Out"]
             else:
-                available_modes = ["Line Out", "Optical Out", "Coax Out", "Bluetooth Out"]
+                available_modes = ["Line Out", "Optical Out", "Coax Out"]
 
             # Convert hardware mode to name
             from pywiim.api.constants import AUDIO_OUTPUT_MODE_MAP